Sharedpref, Role, Messages
Made the server populate the token with the role of user aswell so we can use it for our rendering of screens.

Made the token and role och the user store in sharedpref in the app.

Made the functionality for the create and view messages. Its not good looking but it works.