Functionality to changes status on tasks
Can now change status on tasks to jump up one "level". todo->ongoing->done->delete from db

Added a field to assign the task to a chosen user that you can choose from a dropdown populated by the db.

Added the username to the token so we can use that in frontend aswell.